requirement.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrdict==2.0.0
attrs==18.2.0
bcrypt==3.1.6
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
crypto==1.4.1
cryptography==2.3.1
cytoolz==0.9.0.1
ddt==1.2.0
et-xmlfile==1.0.1
eth-abi==1.2.2
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.3.0
eth-utils==1.3.0
hexbytes==0.1.0
idna==2.7
jdcal==1.4
lru-dict==1.1.6
more-itertools==4.3.0
Naked==0.1.31
openpyxl==2.5.9
paramiko==2.4.2
parsimonious==0.8.1
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.7.0
PyEmail==0.0.1
PyMySQL==0.9.2
PyNaCl==1.3.0
pypiwin32==223
pytest==3.9.2
python-dateutil==2.7.5
pytz==2018.7
pywin32==224
PyYAML==3.13
requests==2.20.0
rlp==1.0.3
selenium==3.14.1
shellescape==3.4.1
six==1.11.0
toolz==0.9.0
urllib3==1.24
web3==4.8.2
websockets==6.0
xlrd==1.1.0

